Fix >4 GiB framed transfers: widen UploadFraming dataLength to UInt64
encode/encodePrefix funneled the data-fork length through UInt32, trapping
on any >4 GiB framed download. forkHeader already writes 64 bits; thread
UInt64 through so it is reachable. Regression test added.

+++ Sources/HeidrunCore/Network/HotlineNetworkClient+Transfers.swift
@@ -203,7 +203,7 @@ extension HotlineNetworkClient {
             creator: creator,
             creationDate: creationDate,
             modificationDate: modificationDate,
-            dataLength: UInt32(content.count),
+            dataLength: UInt64(content.count),
             encoding: stringEncoding
         )
         try await actor.sendBytes(prefix)
